web-animations/interfaces/AnimationEffectTiming/endDelay.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffectTiming.endDelay</title>
<link rel="help" href="https://drafts.csswg.org/web-animations/#dom-animationeffecttiming-enddelay">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(t => {
  const anim = createDiv(t).animate(null, 2000);
  anim.effect.timing.endDelay = 123.45;
  assert_time_equals_literal(anim.effect.timing.endDelay, 123.45,
                             'set endDelay 123.45');
  assert_time_equals_literal(anim.effect.getComputedTiming().endDelay, 123.45,
                             'getComputedTiming() after set endDelay 123.45');
}, 'Allows setting the endDelay to a positive number');

test(t => {
  const anim = createDiv(t).animate(null, 2000);
  anim.effect.timing.endDelay = -1000;
  assert_equals(anim.effect.timing.endDelay, -1000, 'set endDelay -1000');
  assert_equals(anim.effect.getComputedTiming().endDelay, -1000,
                'getComputedTiming() after set endDelay -1000');
}, 'Allows setting the endDelay to a negative number');

test(t => {
  const anim = createDiv(t).animate(null, 2000);
  assert_throws({ name: 'TypeError' }, () => {
    anim.effect.timing.endDelay = Infinity;
  }, 'we can not assign Infinity to timing.endDelay');
}, 'Throws when setting the endDelay to infinity');

test(t => {
  const anim = createDiv(t).animate(null, 2000);
  assert_throws({ name: 'TypeError' }, () => {
    anim.effect.timing.endDelay = -Infinity;
  }, 'we can not assign negative Infinity to timing.endDelay');
}, 'Throws when setting the endDelay to negative infinity');

</script>
</body>
